Module loading refactoring
